Refactor FileLinkHelper into FileSheet, move file utilities to FileHelper, delete FileApp
- Create Views/Sheets/FileSheet.cs as proper ViewBase with Build() and static CreateLinkClickHandler
- Move LanguageMap, ImageExtensions, GetLanguage, IsImageExtension to FileHelper
- Delete FileLinkHelper.cs and FileApp.cs (dead code)
- Update all consumers (7 apps) to use FileSheet directly
- Rename test file to FileSheetTests.cs, remove flaky ConcurrentReads test
