+++ src/Ivy.Tendril/Apps/PullRequestApp.cs
@@ -6,6 +6,7 @@ using Ivy.Tendril.Apps.PullRequest;
 using Ivy.Tendril.Apps.PullRequest.Dialogs;
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
+using Ivy.Tendril.Views.Sheets;
 using Microsoft.Extensions.Logging;
 
 namespace Ivy.Tendril.Apps;
@@ -33,11 +34,7 @@ public class PullRequestApp : ViewBase
                 ? Text.P("Plan not found or empty.")
                 : (object)new Markdown(MarkdownHelper.AnnotateAllBrokenLinks(content, planService.PlansDirectory))
                     .DangerouslyAllowLocalFiles()
-                    .OnLinkClick(FileLinkHelper.CreateFileLinkClickHandler(openFile));
-
-            var repoPaths = plan?.GetEffectiveRepoPaths(config) ?? [];
-            var fileLinkSheet = FileLinkHelper.BuildFileLinkSheet(
-                openFile.Value, () => openFile.Set(null), repoPaths, config);
+                    .OnLinkClick(FileSheet.CreateLinkClickHandler(openFile));
 
             var sheet = new Sheet(
                 () => isOpen.Set(false),
@@ -45,7 +42,7 @@ public class PullRequestApp : ViewBase
                 plan?.Title ?? folderName
             ).Width(Size.Half()).Resizable();
 
-            return fileLinkSheet is not null ? new Fragment(sheet, fileLinkSheet) : sheet;
+            return new Fragment(sheet, new FileSheet(openFile, config));
         });
         var databaseService = UseService<IPlanDatabaseService>();
         var jobService = UseService<IJobService>();
